+++ src/echo/__init__.py
@@ -1,3 +1,3 @@
 """Echo programming language."""
 
-__version__ = "1.1.7"
+__version__ = "1.1.8"